Create basic structure for modalComments.
